# openapi/synapse.yaml
openapi: 3.1.0
info:
  title: Synapse API
  version: 0.1.0
  summary: Organizational Brain service for AI agents and team workflows.
  description: |
    Synapse is the governed layer that gives agents and teams permission-aware
    retrieval, a versioned skill registry, bounded memory, a user/team/org
    context service, a policy-guarded tool/connector (MCP) gateway, and durable
    workflow orchestration.

    Production identity is carried by a verified bearer JWT (JWKS, static RS256,
    or HS256). Trusted X-Principal-Id/X-Tenant-Id headers remain available for
    development or a trusted identity-aware gateway. Every governed endpoint is
    attributable for audit.

  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: ResQ Software
    url: https://github.com/resq-software/synapse
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /skills.register:
    post:
      tags: [skills]
      operationId: skillsRegister
      summary: Register (and version) a skill.
      description: |
        Registers a skill in the registry. A given `(skill_id, version)` pair is
        immutable once registered; new behavior must be published under a new
        version.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Skill'
      responses:
        '200':
          description: Skill registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillRegisterResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /retrieve:
    post:
      tags: [retrieval]
      operationId: retrieve
      summary: Permission-aware hybrid retrieval.
      description: |
        Runs permission-aware retrieval (hybrid vector + lexical, optional
        rerank and graph expansion) scoped by tenant, team, and ACLs. Returns a
        `trace_id` so callers can correlate logs.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrieveRequest'
      responses:
        '200':
          description: Retrieval results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /runs.start:
    post:
      tags: [runs]
      operationId: runsStart
      summary: Start a durable workflow run.
      description: |
        Starts a durable workflow run (sync or async). Run state is persisted so
        it survives restarts; when `callbacks.human_approval` is set, the run
        suspends and returns an opaque `resume_token`.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunsStartRequest'
      responses:
        '200':
          description: Run started (or suspended awaiting approval).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /runs.resume:
    post:
      tags: [runs]
      operationId: runsResume
      summary: Resume a suspended run.
      description: |
        Resumes a suspended run using its opaque `token` (e.g. after human
        approval), applying `resume_input` and continuing the workflow.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunsResumeRequest'
      responses:
        '200':
          description: Run resumed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Verified JWT; required when Synapse is configured for JWT authentication.
  parameters:
    PrincipalId:
      name: X-Principal-Id
      in: header
      required: false
      description: Trusted-header identity for development/gateway mode; ignored when JWT auth is configured.
      schema:
        type: string
      example: alice@resq
    TenantId:
      name: X-Tenant-Id
      in: header
      required: false
      description: Owning tenant/organization for this request.
      schema:
        type: string
      example: resq
    TeamIds:
      name: X-Team-Ids
      in: header
      required: false
      description: Comma-separated team ids the principal is acting within.
      schema:
        type: string
      example: team-ops,team-air
    Role:
      name: X-Role
      in: header
      required: false
      description: Platform/business role of the principal, used for RBAC.
      schema:
        type: string
      example: operator
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid bearer token or trusted principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted by policy.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The referenced resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      description: Stable error envelope returned for all failure responses.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable error code.
              examples: [not_found, bad_request, unauthorized, forbidden, db_error, internal]
            message:
              type: string
              description: Human-readable error message.
...
    Skill:
      type: object
      description: A versioned, governed capability an agent or workflow node can invoke.
      required: [skill_id, version, name]
      properties:
        skill_id:
          type: string
          description: Stable logical identifier, unique within a tenant.
          example: skill.summarize_incident
        version:
          type: string
          description: Semantic version. Immutable per (skill_id, version).
          example: 1.0.0
        name:
          type: string
          description: Human-readable display name.
          example: Summarize Incident
        summary:
          type: string
          description: Short description used for discovery and routing.
        owners:
          type: array
          description: Teams or principals responsible for the skill.
          items:
            type: string
        triggers:
          type: array
          description: Natural-language intents/events that route to this skill.
          items:
            type: string
        input_schema:
          type: object
          description: Nested JSON Schema (draft 2020-12) validating accepted input.
        output_schema:
          type: object
          description: Nested JSON Schema (draft 2020-12) validating produced output.
        required_tools:
          type: array
          description: Tool/connector ids the skill depends on.
          items:
            type: string
        policy_tags:
          type: array
          description: Policy labels evaluated by the Policy & Access Gateway.
          items:
            type: string
        examples:
          type: array
          description: Illustrative input/output pairs for docs and eval sets.
          items:
            type: object
            properties:
              input:
                type: object
              output:
                type: object
...
    SkillRegisterResponse:
      type: object
      description: Result of registering a skill.
      required: [skill_id, version, status]
      properties:
        skill_id:
          type: string
        version:
          type: string
        status:
          type: string
          description: Registration status.
          example: registered
...
    RetrieveScope:
      type: object
      description: Team/project fences layered on top of ACLs.
      properties:
        team_ids:
          type: array
          items:
            type: string
        project_ids:
          type: array
          items:
            type: string
    RetrievalOptions:
      type: object
      description: Retrieval tuning options.
      properties:
        mode:
          type: string
          description: Retrieval mode.
          enum: [hybrid, vector, lexical]
          default: hybrid
        top_k:
          type: integer
          description: Maximum number of results to return.
          minimum: 1
          default: 10
        rerank:
          type: boolean
          description: Whether to apply a second-stage reranker.
          default: false
        include_graph:
          type: boolean
          description: Whether to include knowledge-graph expansion.
          default: false
    RetrieveRequest:
      type: object
      description: A permission-aware retrieval request.
      required: [tenant_id, principal_id, query]
      properties:
        tenant_id:
          type: string
          example: resq
        principal_id:
          type: string
          example: alice@resq
        query:
          type: string
          description: The natural-language (and/or keyword) query.
        scope:
          $ref: '#/components/schemas/RetrieveScope'
        retrieval:
          $ref: '#/components/schemas/RetrievalOptions'
    RetrievalResult:
      type: object
      description: A single retrieval hit.
      required: [chunk_id, doc_id, score, text]
      properties:
        chunk_id:
          type: string
        doc_id:
          type: string
        score:
          type: number
          format: float
          description: Relevance score for the hit.
        text:
          type: string
        source_uri:
          type: string
          description: URI of the source document, when available.
        metadata:
          type: object
          additionalProperties: true
    RetrieveResponse:
      type: object
      description: Retrieval results with a correlation trace id.
      required: [results, trace_id]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/RetrievalResult'
        trace_id:
          type: string
          description: Correlation id for logs/observability.
...
    RunCallbacks:
      type: object
      description: Callback hooks for a durable run.
      properties:
        human_approval:
          type: boolean
          description: If true, the run suspends for human approval and returns a resume token.
          default: false
        webhook:
          type: string
          description: Optional webhook URL notified on run state transitions.
    RunsStartRequest:
      type: object
      description: Request to start a durable workflow run.
      required: [tenant_id, run_type]
      properties:
        tenant_id:
          type: string
          example: resq
        run_type:
          type: string
          description: The type/category of run.
          example: incident_response
        workflow_id:
          type: string
          description: Identifier of the workflow definition to execute.
          example: wf.incident.v1
        input:
          type: object
          description: Initial input for the run.
          additionalProperties: true
        callbacks:
          $ref: '#/components/schemas/RunCallbacks'
    RunsResumeRequest:
      type: object
      description: Request to resume a suspended run.
      required: [run_id, token]
      properties:
        run_id:
          type: string
          description: Identifier of the suspended run.
        token:
          type: string
          description: Opaque resume token issued when the run suspended.
        resume_input:
          type: object
          description: Input applied on resume (e.g. approval decision).
          additionalProperties: true
    RunResponse:
      type: object
      description: State of a run after start/resume.
      required: [run_id, status]
      properties:
        run_id:
          type: string
        status:
          type: string
          description: Current run status (e.g. running, suspended, completed).
        resume_token:
          type: string
          description: Opaque token to resume a suspended (e.g. awaiting-approval) run.
...